+++ plugin/src/Features.ts
@@ -2,12 +2,21 @@ export default interface Features {
   airPlay?: boolean;
   backgroundPlayback?: boolean;
   googleCastSDK?: {
-    android: string | { version: string };
+    appId?: string;
+    messageNamespace?: string;
+    android:
+      | string
+      | {
+          version: string;
+          appId?: string;
+          messageNamespace?: string;
+        };
     ios:
       | string
       | {
           version: string;
           appId?: string;
+          messageNamespace?: string;
           localNetworkUsageDescription?: string;
         };
   };